add by mass

+++ src/main/java/com/tdsi/sn/app_moblile_api/Services/Verification.java
@@ -52,8 +52,8 @@ public class Verification {
         }
         return h;
     }
-    public int getP(int password,int telephone){
-        return password*telephone;
+    public BigInteger getP(int password,int telephone){
+        return BigInteger.valueOf(password*telephone);
     }
     public BigInteger getPu(int password,int telephone){
         BigInteger  g = Verification.g;
